Added emptyRuleSuggester
Requires the full text of the file, so `suggestMutation` now takes in an
`IFileInfo` with both `linesRaw` and `text` instead of just `linesRaw`.

This is also a great example of a rule that intentionally leaves source
code in an "ugly" state. The expected result for the test case has a
bunch of blank lines and trailing whitespace. Those aren't the
responsibility of this rule, so it doesn't fix them. If the user wants
to configure their own `trailing_whitespace` and
`consecutive_blank_lines` rules they can.

Fixes #3.